# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MediaMediaType = typeof MediaMediaType[keyof typeof MediaMediaType];


export const MediaMediaType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface Media {
  id: number;
  media_type: MediaMediaType;
  title: string;
  overview: string;
  /** @nullable */
  poster_path: string | null;
  /** @nullable */
  backdrop_path: string | null;
  /** @nullable */
  release_date: string | null;
  vote_average: number;
  vote_count: number;
  genre_ids: number[];
}

export interface MediaList {
  items: Media[];
  page: number;
  total_pages: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface Season {
  season_number: number;
  name: string;
  episode_count: number;
  /** @nullable */
  air_date: string | null;
}

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profile_path: string | null;
}

export type MediaDetails = Media & ({
  /** @nullable */
  runtime: number | null;
  genres: Genre[];
  seasons: Season[];
  cast: CastMember[];
  production_companies: string[];
});

export interface Person {
  id: number;
  name: string;
  /** @nullable */
  profile_path: string | null;
  known_for_department: string;
}

export interface SearchResults {
  items: Media[];
  people: Person[];
  total_results: number;
}

export interface GenreResponse {
  movies: Genre[];
  tv: Genre[];
}

/**
 * Content not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Catalog service unavailable
 */
export type ServiceUnavailableResponse = ErrorResponse;

export type MediaTypeParameter = typeof MediaTypeParameter[keyof typeof MediaTypeParameter];


export const MediaTypeParameter = {
  movie: 'movie',
  tv: 'tv',
} as const;

export type GetDiscoverMediaParams = {
type: MediaTypeParameter;
category: GetDiscoverMediaCategory;
genre?: number;
};

export type GetDiscoverMediaCategory = typeof GetDiscoverMediaCategory[keyof typeof GetDiscoverMediaCategory];


export const GetDiscoverMediaCategory = {
  popular: 'popular',
  top_rated: 'top_rated',
  now_playing: 'now_playing',
  upcoming: 'upcoming',
} as const;

export type SearchCatalogParams = {
/**
 * @minLength 1
 */
query: string;
filter?: SearchCatalogFilter;
};

export type SearchCatalogFilter = typeof SearchCatalogFilter[keyof typeof SearchCatalogFilter];


export const SearchCatalogFilter = {
  all: 'all',
  movie: 'movie',
  tv: 'tv',
  person: 'person',
} as const;