# lib/api-zod/src/generated/types/castMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profile_path: string | null;
}