# lib/api-zod/src/generated/types/genreResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
import type { Genre } from './genre';

export interface GenreResponse {
  movies: Genre[];
  tv: Genre[];
}