# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaMediaType } from './mediaMediaType';

export interface Media {
  id: number;
  media_type: MediaMediaType;
  title: string;
  overview: string;
  /** @nullable */
  poster_path: string | null;
  /** @nullable */
  backdrop_path: string | null;
  /** @nullable */
  release_date: string | null;
  vote_average: number;
  vote_count: number;
  genre_ids: number[];
}