# lib/api-zod/src/generated/types/mediaDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
import type { CastMember } from './castMember';
import type { Genre } from './genre';
import type { Media } from './media';
import type { Season } from './season';

export type MediaDetails = Media & ({
  /** @nullable */
  runtime: number | null;
  genres: Genre[];
  seasons: Season[];
  cast: CastMember[];
  production_companies: string[];
});