# lib/api-zod/src/generated/types/mediaList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
import type { Media } from './media';

export interface MediaList {
  items: Media[];
  page: number;
  total_pages: number;
}