# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */

export interface Person {
  id: number;
  name: string;
  /** @nullable */
  profile_path: string | null;
  known_for_department: string;
}