# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */
import type { Media } from './media';
import type { Person } from './person';

export interface SearchResults {
  items: Media[];
  people: Person[];
  total_results: number;
}