# lib/api-zod/src/generated/types/season.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cinema Stream API
 * OpenAPI spec version: 0.1.0
 */

export interface Season {
  season_number: number;
  name: string;
  episode_count: number;
  /** @nullable */
  air_date: string | null;
}